+++ relay-server/src/services/store.rs
@@ -506,9 +506,19 @@ impl StoreService {
         &self,
         topic: KafkaTopic,
         // Takes message by value to ensure it is not being produced twice.
-        message: KafkaMessage,
+        mut message: KafkaMessage,
     ) -> Result<(), StoreError> {
         relay_log::trace!("Sending kafka message of type {}", message.variant());
+
+        if let KafkaMessage::Span {
+            ref mut ignore_trace_id_partitioning,
+            ..
+        } = message
+        {
+            let global_config = self.global_config.current();
+            *ignore_trace_id_partitioning =
+                global_config.options.spans_ignore_trace_id_partitioning;
+        }
 
         let topic_name = self.producer.client.send_message(topic, &message)?;
 
@@ -932,6 +942,7 @@ impl StoreService {
                     scoping.project_id.to_string(),
                 )]),
                 message: span,
+                ignore_trace_id_partitioning: false,
             },
         )?;
 
@@ -1588,6 +1599,8 @@ enum KafkaMessage<'a> {
         headers: BTreeMap<String, String>,
         #[serde(flatten)]
         message: SpanKafkaMessage<'a>,
+        #[serde(skip)]
+        ignore_trace_id_partitioning: bool,
     },
     Log {
         #[serde(skip)]
@@ -1631,7 +1644,17 @@ impl Message for KafkaMessage<'_> {
             Self::AttachmentChunk(message) => message.event_id.0,
             Self::UserReport(message) => message.event_id.0,
             Self::ReplayEvent(message) => message.replay_id.0,
-            Self::Span { message, .. } => message.trace_id.0,
+            Self::Span {
+                message,
+                ignore_trace_id_partitioning,
+                ..
+            } => {
+                if *ignore_trace_id_partitioning {
+                    Uuid::nil()
+                } else {
+                    message.trace_id.0
+                }
+            }
 
             // Monitor check-ins use the hinted UUID passed through from the Envelope.
             //