+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Add OTA Updates Event Context for Expo and other applications. ([#4690](https://github.com/getsentry/relay/pull/4690))
 - Add data categories for Seer. ([#4692](https://github.com/getsentry/relay/pull/4692))
 - Allow pii scrubbing of all span `sentry_tags` fields. ([#4698](https://github.com/getsentry/relay/pull/4698))
+- Add killswitch for trace id partitioning. ([#4706](https://github.com/getsentry/relay/pull/4706))
 - Add OTA Updates Event Context for Replay Events. ([#4711](https://github.com/getsentry/relay/pull/4711))
 
 **Bug Fixes**: